# openapi.yaml
openapi: 3.0.0
info:
  title: Mini Task API
  description: RESTful API for Task Management System with Authentication, Authorization, Rate Limiting & Idempotency
  version: 1.0.0
  contact:
    name: API Support
    email: support@minitaskapi.com
  license:
    name: ISC
servers:
  - url: http://localhost:3000/api/v1
    description: Development Server
  - url: https://api.example.com/api/v1
    description: Production Server
...
paths:
...
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get new access token using refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Access token refreshed successfully
                  accessToken:
                    type: string
        '401':
          description: Refresh token blacklisted (user logged out)
        '403':
          description: Invalid or expired refresh token
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from login
...